requirements.txt

pyspark==3.4.1

kafka-python==2.0.2

nltk==3.8.1

scikit-learn==1.2.2
joblib==1.3.1

numpy==1.23.5

scipy
regex
emot
pyvi
scipy
IPython
confluent_kafka
openpyxl